requirements-production.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
gunicorn==21.2.0

# Database & Storage
sqlalchemy==2.0.23
pyodbc==5.0.1
azure-storage-blob==12.19.0
azure-identity==1.15.0

# API & Web
requests==2.31.0
pandas==2.0.3
plotly==5.17.0
pydantic==2.5.0

# Configuration & Security
python-dotenv==1.0.0
azure-keyvault-secrets==4.7.0

# Lightweight AI (free tier)
scikit-learn==1.3.0
numpy==1.24.3
